backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "gunicorn>=23.0.0",
    "imgviz>=1.7.6",
    "labelme>=5.8.1",
    "opencv-python>=4.11.0.86",
    "pika>=1.3.2",
    "pycocotools>=2.0.8",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "torch>=2.6.0",
    "ultralytics>=8.3.96",
    "uvicorn[standard]>=0.34.0",
    "redis",
    "numba>=0.61.0",
    "fastremap>=1.15.1",
    "roifile>=2025.2.20",
    "openpyxl>=3.1.5",
    "cellpose>=3.1.1.1",
    "shapely>=2.0.7",
    "numpy>=2.0.2",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
